Update say.py
To start...I added in a try and except for having the lack of perms of the bot deleting your message.
Another addition is a try and except for an empty command([p]say )
Both of these errors return a message to the user that initiated the command and let's them know what happened. In a embed for a clean look.
Only players who have admin permissions can use this command, because people would abuse this power.