Added distance cog
Glad to announce, another V3 cog is here! This is relatively for Pokemon Go, you entering a set of coords and getting the distance. It is not EXACT but is close. Also get a cooldown timer which is from the Pokemon Go site. Is known to work locally and on a server.